+++ src/main/java/com/elmakers/mine/bukkit/plugins/magic/Wand.java
@@ -11,6 +11,7 @@ import java.util.Map;
 import java.util.Map.Entry;
 
 import org.apache.commons.lang.StringUtils;
+import org.bukkit.ChatColor;
 import org.bukkit.Material;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.Inventory;
@@ -211,7 +212,7 @@ public class Wand {
 			materialName = " : " + materialName;
  		}
 		if (spell != null) {
-			setName(spell.getName() + materialName + " (" + baseName + ")");
+			setName(ChatColor.RED + spell.getName() + ChatColor.GRAY + materialName + ChatColor.WHITE + " (" + baseName + ")");
 		} else {
 			setName(getBaseName());
 		}