+++ CHANGELOG.md
@@ -4,6 +4,8 @@
  - Allow use of /wand list and /spells from the console
  - Colorize active spell text
  - Fix cycling materials if you use slot 8 to hold the wand
+ - Add "cat" spell (OcelotSpell)
+ - Add Head Shrinker spell
  
 ## 2.2.3